+++ src/types/tag.rs
@@ -2,6 +2,7 @@ use crate::types::HeadingLevel;
 
 #[derive(Debug, PartialEq)]
 pub enum Tag<'a> {
+    BlockQuote,
     FencedCodeBlock(&'a str),
     Heading(HeadingLevel),
     IndentedCodeBlock,