+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Support HTML blocks.
+- Support block quotes.
 
 ## 0.6.0 - 2021-11-28